Abstracted GrammarTreeNodeObject to an interface
