+++ Assets/Code/IGrammarTreeNodeObject.cs
@@ -1,0 +1,12 @@
+﻿public interface IGrammarTreeNodeObject
+{
+    GrammarTreeNode Node { get; }
+
+    string Contents();
+
+    GrammarNodeType NodeType();
+
+    void SetNode(GrammarTreeNode node);
+
+    void UpdateDisplay();
+}
